+++ src/Filters/FiltersExact.php
@@ -3,6 +3,7 @@
 namespace Spatie\QueryBuilder\Filters;
 
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Database\Eloquent\Relations\Relation;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Str;
 
@@ -47,11 +48,13 @@ class FiltersExact implements Filter
             return false;
         }
 
-        if (Str::startsWith($property, $query->getModel()->getTable().'.')) {
+        $firstRelationship = Str::camel(explode('.', $property)[0]);
+
+        if (! method_exists($query->getModel(), $firstRelationship)) {
             return false;
         }
 
-        return true;
+        return is_a($query->getModel()->{$firstRelationship}(), Relation::class);
     }
 
     protected function withRelationConstraint(Builder $query, $value, string $property)
@@ -64,8 +67,8 @@ class FiltersExact implements Filter
                 ];
             });
 
-        $query->whereHas($relation, function (Builder $query) use ($value, $relation, $property) {
-            $this->relationConstraints[] = $property = $query->getModel()->getTable().'.'.$property;
+        $query->whereHas($relation, function (Builder $query) use ($value , $property) {
+            $this->relationConstraints[] = $property = $query->qualifyColumn($property);
 
             $this->__invoke($query, $value, $property);
         });